+++ src/server/components/character/carriable.ts
@@ -2,8 +2,8 @@ import { Component, Components } from "@flamework/components";
 import { OnStart } from "@flamework/core";
 import { Players, Workspace } from "@rbxts/services";
 import { Inject } from "shared/inject";
+import { Character } from ".";
 import { KeyInteractable } from "../interactable/key-interactable";
-import { Character } from "./character";
 import { RagdollServer } from "./ragdoll-server";
 
 interface Attributes {
